Added version 5.6.2867

+++ chromium/ash/services/secure_channel/ble_characteristics_finder.cc
@@ -1,13 +1,13 @@
-// Copyright 2015 The Chromium Authors. All rights reserved.
+// Copyright 2015 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "ash/services/secure_channel/ble_characteristics_finder.h"
 
-#include "ash/components/multidevice/logging/logging.h"
 #include "ash/services/secure_channel/background_eid_generator.h"
 #include "base/bind.h"
 #include "base/strings/string_util.h"
+#include "chromeos/ash/components/multidevice/logging/logging.h"
 #include "device/bluetooth/bluetooth_adapter.h"
 #include "device/bluetooth/bluetooth_device.h"
 #include "device/bluetooth/bluetooth_remote_gatt_characteristic.h"
@@ -184,7 +184,7 @@ void BluetoothLowEnergyCharacteristicsFinder::OnRemoteCharacteristicRead(
 
   if (error_code.has_value()) {
     PA_LOG(ERROR) << "OnWriteRemoteCharacteristicError() Error code: "
-                  << error_code.value();
+                  << static_cast<int>(error_code.value());
     service_ids_pending_eid_read_.erase(service_id);
     if (!has_callback_been_invoked_)
       NotifyFailureIfNoPendingEidCharReads();
