+++ chromium/chrome/browser/ash/login/screens/recommend_apps_screen.h
@@ -1,4 +1,4 @@
-// Copyright 2018 The Chromium Authors. All rights reserved.
+// Copyright 2018 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -94,7 +94,7 @@ class RecommendAppsScreen : public BaseScreen,
   // Skip the screen for testing if set to true.
   bool skip_for_testing_ = false;
 
-  int recommended_app_count_ = 0;
+  size_t recommended_app_count_ = 0;
 
   base::raw_ptr<PrefService> pref_service_;
