+++ chromium/chrome/browser/ui/webui/chromeos/login/check_passwords_against_cryptohome_helper.cc
@@ -1,22 +1,22 @@
-// Copyright 2022 The Chromium Authors. All rights reserved.
+// Copyright 2022 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "chrome/browser/ui/webui/chromeos/login/check_passwords_against_cryptohome_helper.h"
 
-#include "ash/components/login/auth/auth_status_consumer.h"
-#include "ash/components/login/auth/cryptohome_authenticator.h"
-#include "ash/components/login/auth/extended_authenticator.h"
-#include "ash/components/login/auth/public/cryptohome_key_constants.h"
-#include "ash/components/login/auth/public/user_context.h"
 #include "ash/constants/ash_features.h"
 #include "base/values.h"
+#include "chromeos/ash/components/login/auth/auth_status_consumer.h"
+#include "chromeos/ash/components/login/auth/cryptohome_authenticator.h"
+#include "chromeos/ash/components/login/auth/extended_authenticator.h"
+#include "chromeos/ash/components/login/auth/public/cryptohome_key_constants.h"
+#include "chromeos/ash/components/login/auth/public/user_context.h"
 
 namespace ash {
 namespace {
 
 // Maximum number of attempts to check scraped password against cryptohome.
-const int kMaximumNumberOfAttempts = 5;
+const size_t kMaximumNumberOfAttempts = 5;
 
 void SetKeyForUserContext(UserContext& user_context,
                           const std::string& password) {
@@ -41,7 +41,7 @@ CheckPasswordsAgainstCryptohomeHelper::CheckPasswordsAgainstCryptohomeHelper(
           on_check_passwords_against_cryptohome_helper_failure_callback)),
       on_check_passwords_against_cryptohome_helper_success_callback_(std::move(
           on_check_passwords_against_cryptohome_helper_success_callback)) {
-  current_password_index_ = 0;
+  current_password_index_ = 0u;
   extended_authenticator_ = ExtendedAuthenticator::Create(this);
   SetKeyForUserContext(user_context_,
                        scraped_passwords_[current_password_index_]);
@@ -57,7 +57,7 @@ void CheckPasswordsAgainstCryptohomeHelper::OnAuthFailure(
   current_password_index_++;
   if (current_password_index_ == kMaximumNumberOfAttempts ||
       current_password_index_ == scraped_passwords_.size()) {
-    current_password_index_ = 0;
+    current_password_index_ = 0u;
     std::move(on_check_passwords_against_cryptohome_helper_failure_callback_)
         .Run();
     return;
